+++ src/packages/tyhUi/packages/textarea/index.js
@@ -1,4 +1,4 @@
-import T from './TyhTextarea.vue'
+import T from './src/TyhTextarea.vue'
 
 T.install = _ => _.component('TyhTextarea', T)